Add formatter and checkstyle
Adding maven plugin to format the java code automatically on builds. The
style of the formatter is opinionated favouring googles formatting. The
benefit of this is it allows us to easily plugin in googles check style
file. The checkstyle.xml has been made more lenient to ignore things
like missing javadoc and abbreviations.